hello-python/hello-python/hello-python/requirements.txt
ldclient-py==3.0.3
